Fix backend implementation
Refactor the backend code to properly implement NestJS and update configurations.

* **Configuration Changes**
  - Remove `ConfigModule` and `ConfigService` imports from `config.ts`.
  - Remove `AppConfigModule` class and export the `config` object.
  - Add `ConfigModule` import to `resume.module.ts` and include it in the `imports` array.
  - Add `ConfigService` import to `resume.service.ts` and include it in the constructor.
  - Get `GOOGLE_GENAI_API_KEY` from `ConfigService`.

* **Controller and Service Changes**
  - Remove `GenerateResumeDto`, `RefineResumeDto`, and `DownloadResumeDto` imports from `resume.controller.ts`.
  - Remove `@ApiBody` decorators from `resume.controller.ts`.
  - Update `generateResume`, `refineResume`, and `downloadResume` methods in `resume.controller.ts` to use direct parameters instead of DTOs.

* **File Deletions**
  - Delete `app.ts`, `controllers/resumeController.ts`, `main.ts`, `middlewares/uploadMiddleware.ts`, `resume/dto/download-resume.dto.ts`, `resume/dto/generate-resume.dto.ts`, `resume/dto/refine-resume.dto.ts`, and `services/resumeService.ts`.

* **Index File Changes**
  - Remove `config` import from `index.ts`.
  - Add `ConfigService` import and get the `ConfigService` instance from the app.
  - Get the port from `ConfigService`.

* **TypeScript Configuration**
  - Add `experimentalDecorators` and `emitDecoratorMetadata` to `compilerOptions` in `tsconfig.json`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/suhailsalim/resume-tailor?shareId=XXXX-XXXX-XXXX-XXXX).
