--- backend/src/app.ts
@@ -1,14 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
-import { ResumeModule } from './resume/resume.module';
-
-@Module({
-  imports: [
-    ConfigModule.forRoot({
-      isGlobal: true,
-      envFilePath: ['.env'],
-    }),
-    ResumeModule,
-  ],
-})
-export class AppModule {}
