+++ backend/src/index.ts
@@ -1,10 +1,11 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { config } from './config/config';
+import { ConfigService } from '@nestjs/config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  const port = config.port;
+  const configService = app.get(ConfigService);
+  const port = configService.get<number>('PORT') || 3001;
   await app.listen(port);
   console.log(`Server running on port ${port}`);
 }
