--- backend/src/main.ts
@@ -1,23 +1,0 @@
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-import { ConfigService } from '@nestjs/config';
-import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
-
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  const configService = app.get(ConfigService);
-  const port = configService.get<number>('PORT') || 3001;
-
-  const config = new DocumentBuilder()
-    .setTitle('Resume Tailor API')
-    .setDescription('API documentation for Resume Tailor')
-    .setVersion('1.0')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document);
-
-  await app.listen(port);
-  console.log(`Server running on port ${port}`);
-}
-
-bootstrap();
