--- backend/src/middlewares/uploadMiddleware.ts
@@ -1,52 +1,0 @@
-import { Injectable, NestMiddleware } from '@nestjs/common';
-import { MulterModule } from '@nestjs/platform-express';
-import { Request, Response, NextFunction } from 'express';
-import multer from 'multer';
-
-const storage = multer.memoryStorage();
-
-const upload = multer({
-  storage,
-  limits: { fileSize: 10 * 1024 * 1024 }, // 10MB max file size
-  fileFilter: (req: Request, file, cb) => {
-    const allowedTypes = [
-      'application/pdf',
-      'application/msword',
-      'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-    ];
-    if (allowedTypes.includes(file.mimetype)) {
-      cb(null, true);
-    } else {
-      cb(new Error('Invalid file type. Only PDF, DOC, and DOCX files are allowed.') as any);
-    }
-  },
-});
-
-@Injectable()
-export class UploadMiddleware implements NestMiddleware {
-  use(req: Request, res: Response, next: NextFunction) {
-    upload.single('resume')(req, res, (err) => {
-      if (err) {
-        return res.status(400).send({ message: err.message });
-      }
-      next();
-    });
-  }
-}
-
-export const UploadModule = MulterModule.register({
-  storage,
-  limits: { fileSize: 10 * 1024 * 1024 },
-  fileFilter: (req: Request, file, cb) => {
-    const allowedTypes = [
-      'application/pdf',
-      'application/msword',
-      'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-    ];
-    if (allowedTypes.includes(file.mimetype)) {
-      cb(null, true);
-    } else {
-      cb(new Error('Invalid file type. Only PDF, DOC, and DOCX files are allowed.') as any);
-    }
-  },
-});
