--- backend/src/resume/dto/download-resume.dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class DownloadResumeDto {
-  @ApiProperty()
-  markdown: string;
-
-  @ApiProperty()
-  format: string;
-}
