--- backend/src/resume/dto/generate-resume.dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class GenerateResumeDto {
-  @ApiProperty({ type: 'string', format: 'binary' })
-  file: any;
-
-  @ApiProperty()
-  jobDescription: string;
-}
