--- backend/src/resume/dto/refine-resume.dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-
-export class RefineResumeDto {
-  @ApiProperty()
-  originalResume: string;
-
-  @ApiProperty()
-  feedback: string;
-}
